:bug: update userUpdate, work on refresh_token

+++ app/datamapper/category.js
@@ -24,11 +24,15 @@ async function findCategoryByUser(userId) {
         values: [userId],
     };
     const result = await pool.query(queryPrepared);
-    return result.rows[0];
+    return result;
 }
 
+
 async function updateData(categories, userId) {
-    await pool.query(`DELETE FROM public.user_like_category WHERE user_id = $1;`, [userId]);
+
+    const findUserCategories = await pool.query(`SELECT category_id FROM public.user_like_category WHERE user_id = $1;`, [userId])
+
+    if (findUserCategories.rowCount !== 0) await pool.query(`DELETE FROM public.user_like_category WHERE user_id = $1;`, [userId]);
 
     for await (const category of categories) {
         pool.query(`INSERT INTO public.user_like_category
